# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gift Multiverse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gift-cards:
    get:
      operationId: getGiftCards
      tags: [giftCards]
      summary: List gift cards with filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: subcategoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: sort
          in: query
          schema:
            type: ["string", "null"]
            enum: [price_asc, price_desc, discount_desc, newest, null]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Gift cards list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftCardListResponse"
...
  /gift-cards/{id}:
    get:
      operationId: getGiftCard
      tags: [giftCards]
      summary: Get gift card by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gift card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftCard"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── ORDERS ────────────────────────────────────────────────────────────────
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get current user's dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── WALLET ────────────────────────────────────────────────────────────────
  /wallet/transactions:
    get:
      operationId: getWalletTransactions
      tags: [wallet]
      summary: Get wallet transactions for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── BLOG ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    GiftCard:
      type: object
      required: [id, brand, name, denomination, originalPrice, sellPrice, discountPct, stock, isActive, subcategoryId, createdAt]
      properties:
        id:
          type: integer
        brand:
          type: string
        name:
          type: string
        denomination:
          type: number
        originalPrice:
          type: number
        sellPrice:
          type: number
        discountPct:
          type: number
        stock:
          type: integer
        placeholderColor:
          type: ["string", "null"]
        isActive:
          type: boolean
        subcategoryId:
          type: integer
        subcategoryName:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    GiftCardListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GiftCard"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    DashboardStats:
      type: object
      required: [totalOrders, pendingOrders, completedOrders, totalSellRequests, pendingSellRequests, walletBalance]
      properties:
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        completedOrders:
          type: integer
        totalSellRequests:
          type: integer
        pendingSellRequests:
          type: integer
        walletBalance:
          type: number
...
    Transaction:
      type: object
      required: [id, userId, type, amount, description, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: number
        description:
          type: string
        ref:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ── BLOG SCHEMAS ──────────────────────────────────────────────────────
...